Populate Board/Game and refactor Player
1. Include important methods in Board and Game classes
2. Create a module for human choice making
